+++ SquirrelDefender/libraries/math/interpolate.cpp
@@ -41,7 +41,7 @@ float get_float_index(float input, const float *array, int max_idx, bool is_asce
     {
         if (input <= array[0])
         {
-            return 0.0f; // Input is less than or equal to the first element (ascending)
+            return (float)0.0; // Input is less than or equal to the first element (ascending)
         }
         if (input >= array[max_idx - 1])
         {
@@ -52,7 +52,7 @@ float get_float_index(float input, const float *array, int max_idx, bool is_asce
     {
         if (input >= array[0])
         {
-            return 0.0f; // Input is greater than or equal to the first element (descending)
+            return (float)0.0; // Input is greater than or equal to the first element (descending)
         }
         if (input <= array[max_idx - 1])
         {
